Summary of Changes:
The most significant change in the Dowdian.DNN.VaultApi solution is the removal of the "Website" project and its associated "SolutionItems" section, which included a placeholder.txt file. No new code or projects have been added to the solution.

List of Changes:

1. Removal of "Website" Project: The "Website" project has been removed from the Dowdian.DNN.VaultApi solution. This is a significant change as it indicates that the project is no longer part of the solution.

2. Removal of "SolutionItems" Section: Along with the "Website" project, its associated "SolutionItems" section, which included a placeholder.txt file, has also been removed. This suggests that the items that were previously part of this section are no longer required in the solution.

3. No Addition of New Code or Projects: There have been no additions of new code or projects to the Dowdian.DNN.VaultApi solution. This indicates that the solution has not been expanded or updated with new functionalities.

References to Code Changes:

The changes are indicated by the lines starting with '-' in the code, which signifies the removal of code or projects. The absence of lines starting with '+' indicates that no new code or projects have been added.